+++ src/posParallelRcpp.cpp
@@ -1,9 +1,10 @@
-// [[Rcpp::depends(RcppParallel)]]
 // [[Rcpp::plugins(cpp11)]]
+// [[Rcpp::depends(BH, RcppParallel)]]
 
 #include <Rcpp.h>
 #include <RcppParallel.h>
-#include <mecab.h>
+#include <boost/algorithm/string.hpp>
+#include "mecab.h"
 
 using namespace Rcpp;
 
@@ -32,8 +33,9 @@ struct TextParseJoin
           ;
         else {
           std::string parsed_morph = std::string(node->surface).substr(0, node->length);
-          std::string feature_result = node->feature;
-          parsed.push_back(parsed_morph + "/" + feature_result.substr(0, feature_result.find(",")));
+          std::vector<std::string> features;
+          boost::split(features, node->feature, boost::is_any_of(","));
+          parsed.push_back(parsed_morph + "/" + features[0]);
         }
       }
 
@@ -76,9 +78,10 @@ struct TextParse
         else {
           std::string parsed_morph = std::string(node->surface).substr(0, node->length);
           // add join mechanism
-          std::string feature_result = node->feature;
+          std::vector<std::string> features;
+          boost::split(features, node->feature, boost::is_any_of(","));
           parsed.push_back(parsed_morph);
-          parsed.push_back(feature_result.substr(0, feature_result.find(",")));
+          parsed.push_back(features[0]);
         }
       }
 
@@ -95,12 +98,22 @@ struct TextParse
 };
 
 // [[Rcpp::export]]
-List posParallelJoinRcpp( std::vector<std::string> text, std::string dict ) {
+List posParallelJoinRcpp( std::vector<std::string> text, std::string sys_dic, std::string user_dic ) {
 
   std::vector< std::vector < std::string > > results(text.size());
   List result;
 
-  mecab_model_t* model = mecab_model_new2(dict.c_str());
+  char arg0[] = "-d";
+  char arg1[sys_dic.length() + 1];
+  strcpy(arg1, sys_dic.c_str());
+  char arg2[] = "-u";
+  char arg3[user_dic.length() + 1];
+  strcpy(arg3, user_dic.c_str());
+  char* argv_model[] = { &arg0[0], &arg1[0], &arg2[0], &arg3[0], NULL };
+  int argc_model = (int)(sizeof(argv_model) / sizeof(argv_model[0])) - 1;
+
+  // Create MeCab model
+  mecab_model_t* model = mecab_model_new(argc_model, argv_model);
   if (!model) {
     Rcerr << "model is NULL" << std::endl;
     return R_NilValue;
@@ -129,12 +142,22 @@ List posParallelJoinRcpp( std::vector<std::string> text, std::string dict ) {
 }
 
 // [[Rcpp::export]]
-List posParallelRcpp( std::vector<std::string> text, std::string dict ) {
+List posParallelRcpp( std::vector<std::string> text, std::string sys_dic, std::string user_dic ) {
 
   std::vector< std::vector < std::string > > results(text.size());
   List result;
 
-  mecab_model_t* model = mecab_model_new2(dict.c_str());
+  char arg0[] = "-d";
+  char arg1[sys_dic.length() + 1];
+  strcpy(arg1, sys_dic.c_str());
+  char arg2[] = "-u";
+  char arg3[user_dic.length() + 1];
+  strcpy(arg3, user_dic.c_str());
+  char* argv_model[] = { &arg0[0], &arg1[0], &arg2[0], &arg3[0], NULL };
+  int argc_model = (int)(sizeof(argv_model) / sizeof(argv_model[0])) - 1;
+
+  // Create MeCab model
+  mecab_model_t* model = mecab_model_new(argc_model, argv_model);
   if (!model) {
     Rcerr << "model is NULL" << std::endl;
     return R_NilValue;